feat(module:input): move input group tests to correct spec file
